結合テストのファイルの移動
